Refactor VCS layers into top-level modules
- Move VcsDriver and VcsProcess out of nested service/layer paths
- Update server, workspace, orchestration, and tests to use new exports
- Rename layer exports to `layer` for consistency
