+++ apps/server/src/orchestration/Layers/ProviderCommandReactor.test.ts
@@ -51,7 +51,7 @@ import { OrchestrationEngineService } from "../Services/OrchestrationEngine.ts";
 import { ProviderCommandReactor } from "../Services/ProviderCommandReactor.ts";
 import * as NodeServices from "@effect/platform-node/NodeServices";
 import { ServerSettingsService } from "../../serverSettings.ts";
-import { VcsDriver, type VcsDriverShape } from "../../vcs/Services/VcsDriver.ts";
+import { VcsDriver, type VcsDriverShape } from "../../vcs/VcsDriver.ts";
 
 const asProjectId = (value: string): ProjectId => ProjectId.make(value);
 const asApprovalRequestId = (value: string): ApprovalRequestId => ApprovalRequestId.make(value);
