+++ apps/server/src/server.test.ts
@@ -104,7 +104,7 @@ import {
 import { WorkspaceEntriesLive } from "./workspace/Layers/WorkspaceEntries.ts";
 import { WorkspaceFileSystemLive } from "./workspace/Layers/WorkspaceFileSystem.ts";
 import { WorkspacePathsLive } from "./workspace/Layers/WorkspacePaths.ts";
-import { VcsDriver, type VcsDriverShape } from "./vcs/Services/VcsDriver.ts";
+import { VcsDriver, type VcsDriverShape } from "./vcs/VcsDriver.ts";
 import { ServerSecretStoreLive } from "./auth/Layers/ServerSecretStore.ts";
 import { ServerAuthLive } from "./auth/Layers/ServerAuth.ts";
