--- apps/server/src/vcs/Services/VcsProcess.ts
@@ -1,32 +1,0 @@
-import { Context, type Effect } from "effect";
-
-import type { VcsError } from "@t3tools/contracts";
-
-export interface VcsProcessInput {
-  readonly operation: string;
-  readonly command: string;
-  readonly args: ReadonlyArray<string>;
-  readonly cwd: string;
-  readonly stdin?: string;
-  readonly env?: NodeJS.ProcessEnv;
-  readonly allowNonZeroExit?: boolean;
-  readonly timeoutMs?: number;
-  readonly maxOutputBytes?: number;
-  readonly truncateOutputAtMaxBytes?: boolean;
-}
-
-export interface VcsProcessOutput {
-  readonly exitCode: number;
-  readonly stdout: string;
-  readonly stderr: string;
-  readonly stdoutTruncated: boolean;
-  readonly stderrTruncated: boolean;
-}
-
-export interface VcsProcessShape {
-  readonly run: (input: VcsProcessInput) => Effect.Effect<VcsProcessOutput, VcsError>;
-}
-
-export class VcsProcess extends Context.Service<VcsProcess, VcsProcessShape>()(
-  "t3/vcs/Services/VcsProcess",
-) {}
